add script options as global variables in tests
